NP-380 release script
