+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@citizensadvice/design-system",
-    "version": "0.11.1-alpha.0",
+    "version": "0.11.0",
     "description": "Citizens Advice Design System",
     "files": [
         "scss",
